Refactored worker controller code into helper to make it more modular
