+++ tests/testSetup.js
@@ -19,38 +19,11 @@
  */
 
 if (process.env.NODE_ENV === 'development') {
-  require('dotenv').config({ path: '../env/development.env' });
+  require('dotenv').config({ path: './env/development.env' });
 } else if (process.env.NODE_ENV === 'production') {
-  require('dotenv').config({ path: '../env/production.env' });
+  require('dotenv').config({ path: './env/production.env' });
 }
 
 // Register babel so that it will transpile ES6 to ES5
 // before our tests run.
 require('babel-register')();
-
-// Disable webpack-specific features for tests since
-// Mocha doesn't know what to do with them.
-require.extensions['.css'] = function () {return null;};
-require.extensions['.png'] = function () {return null;};
-require.extensions['.jpg'] = function () {return null;};
-
-// Configure JSDOM and set global variables
-// to simulate a browser environment for tests.
-var jsdom = require('jsdom').jsdom;
-
-var exposedProperties = ['window', 'navigator', 'document'];
-
-global.document = jsdom('');
-global.window = document.defaultView;
-Object.keys(document.defaultView).forEach((property) => {
-  if (typeof global[property] === 'undefined') {
-    exposedProperties.push(property);
-    global[property] = document.defaultView[property];
-  }
-});
-
-global.navigator = {
-  userAgent: 'node.js'
-};
-
-documentRef = document;  //eslint-disable-line no-undef
