+++ tests/worker.spec.js
@@ -1,45 +1,22 @@
 /* eslint-disable */
 
+// Tools to check out - Winston Logger
+
 // Dependencies
 const expect = require('chai').expect;
 const sinon = require('sinon');
 
 // Modules
+const worker = require('../worker/worker.js');
 const workerHandler = require('../worker/worker_controller.js');
+const helpers = require('../worker/helper.js');
 const script = require('../worker/scripts/scenario');
 
 describe('Worker', () => {
-  describe('handleJob function', () => {
-    // Worker executes job with the provided "script"
-    it('should perform a POST request to master when spun up', (done) => {
-
-    });
-
-    it('should execute job with provided script', sinon.test((done) => {
-      const testScript = this.spy(script);
-      workerhandler.handleJob(100, testScript);
-
-      testScript.called();
-
-    }));
-    // Worker executes all the jobs it was provided
-    it('should post to the end point provided to ask for work', sinon.test((done) => {
-      // const checkCompletion = this.spy(workerHandler, 'handleJob');
-      // const testJob = [ 1, 2, 3 ];
-      // // Feed in jobs number
-      // checkCompletion(testJob);
-      // // Check if results.length is equal to jobs.length
-
-
-      // done();
-    }));
-    // Results are POSTed to a particular address
-    it('should POST results to the provided IP address', () => {
-
-    });
-    // Worker POSTs for more work after completion
-    it('should ask for more work after completion', () => {
-
+  describe('saveActionResultsToDB', () => {
+    it('should save ', (done) => {
+      // 
+      done();
     });
   });
 });
