+++ worker/worker_controller.js
@@ -1,8 +1,7 @@
 // Dependencies
 const request = require('request');
 const scenariorunner = require('./scripts/scenario');
-const Action = require('../models/ActionsModel');
-const Spawn = require('../models/SpawnsModel');
+const helpers = require('./helper');
 
 // Global Variable
 let jobsCompleted = 0;
@@ -40,35 +39,12 @@ const handleJob = jobs => {
     */
       // For each job result, save actions to the actions database
       const actionsResults = runresults.transactionTimes;
-      for (let i = 0; i < actionsResults.length; i++) {
-        const actionData = {
-          statusCode: actionsResults[i].statusCode,
-          elapsedTime: actionsResults[i].elapsedTime,
-          id_scenario: job.scenarioID,
-        };
-        const newAction = new Action(actionData);
-        newAction.save()
-          .then(() => {
-            console.log('Successfully saved');
-          })
-          .catch(err => {
-            console.error(err);
-          });
-      }
+      helpers.saveActionResultsToDB(actionsResults, job);
 
       // Save to spawn database
-      const spawnData = {
-        totalTime: runresults.scenarioTime,
-        id_scenario: job.scenarioID,
-      };
-      const newSpawn = new Spawn(spawnData);
-      newSpawn.save()
-        .then(() => {
-          console.log('Successfully saved');
-        })
-        .catch(err => {
-          console.error(err);
-        });
+      helpers.saveSpawnsToDB(runresults, job);
+
+      // Add to results
       results.push(runresults);
       jobsCompleted++;
     });
@@ -82,17 +58,7 @@ const handleJob = jobs => {
   });
 
   // Request more work from master
-  request.post(requestJob, (error, response, body) => {
-    if (error) {
-      console.error(error);
-    } else if (body === 'done') {
-      console.log('Jobs completed is ', jobsCompleted);
-      process.exit();
-    } else {
-      // Recursively ask for more work if available
-      handleJob(JSON.parse(body).job);
-    }
-  });
+  request.post(requestJob, helpers.responseFromMasterCallback);
 };
 
-module.exports = { handleJob };
+module.exports = { handleJob, jobsCompleted };
